auth-15 - making authorized requests

+++ src/components/WorkoutDetails.tsx
@@ -1,6 +1,7 @@
 import { Workout } from '../types/types'
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext';
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
+import { useAuthContext } from '../hooks/useAuthContext';
 
 type WorkoutDetailsProps = {
     workout: Workout
@@ -8,9 +9,20 @@ type WorkoutDetailsProps = {
 
 const WorkoutDetails = ({ workout }: WorkoutDetailsProps) => {
     const { dispatch } = useWorkoutsContext();
+    const userState = useAuthContext()
+    const user = userState.state.user;
     const handleClick = async () => {
+
+        if (!user) {
+            return;
+        }
+
         const response = await fetch(`/api/workouts/${workout._id}`, {
             method: 'DELETE',
+            headers: {
+                'Content-Type': 'application/json',
+                'Authorization': `Bearer ${user.token}`,
+            }
         })
         const json = await response.json()
         if (response.ok) {
